feat(calendar): implement caching for event creation, update, and deletion
- Added `_mirror_to_cache` and `_forget_from_cache` functions to handle event caching.
- Updated `create_event` and `patch_event` to cache events after creation and updates.
- Implemented cache invalidation in `delete_event`.

feat(gmail): enhance HTML to Markdown conversion

- Updated `_html_to_text` to `_html_to_markdown` for better Markdown support.
- Added support for various HTML elements including headings, lists, and links.
- Implemented tests for Markdown conversion.

feat(kotx): add caching for kotx tasks

- Introduced `cache_kotx_task` to fetch and cache kotx task documents.
- Combined task details into a single document for embedding.

fix(points): change points amount to integer

- Updated points adjustment to return an integer instead of a float.
- Adjusted database schema to enforce integer type for points.

feat(notes): add generated tsvector column for hybrid search

- Implemented a generated `tsv` column for notes to support hybrid search.
- Created GIN index for efficient text search.

test(tests): add comprehensive tests for document caching and email preprocessing

- Added tests for calendar event content and metadata handling.
- Implemented tests for kotx task content generation.
- Created tests for HTML to Markdown conversion in email preprocessing.

+++ backend/app/api/points.py
@@ -33,7 +33,7 @@ router = APIRouter(prefix="/points", tags=["points"])
 
 
 class TotalRead(BaseModel):
-    total: float
+    total: int
 
 
 class AdjustPayload(BaseModel):
@@ -44,7 +44,7 @@ class AdjustPayload(BaseModel):
 
 class PointsLogEntryRead(BaseModel):
     id: uuid.UUID
-    amount: float
+    amount: int
     source: str
     reason: str
     caller: str | None
